Add getDiagnosticProperty method
