--- language-server/modules/langserver-commons/src/main/java/org/ballerinalang/langserver/commons/codeaction/spi/DiagnosticPropertyKey.java
@@ -1,9 +1,0 @@
-package org.ballerinalang.langserver.commons.codeaction.spi;
-
-/**
- * Diagnostic Property Names.
- */
-public enum DiagnosticPropertyKey {
-    DIAG_PROP_INCOMPATIBLE_TYPES_EXPECTED,
-    DIAG_PROP_INCOMPATIBLE_TYPES_FOUND
-}
